+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## Changed
+
+- Made actor handling more strict. Now referencing a user/group that doesn't exist explicitly in rules will cause an error, including a user that doesn't exist in a user revoke policy (`revokeUsers`) will cause an error, and attempting to grant permissions to a user outside the scope of the user revoke policy will cause an error.
+
+- Exclude default postgres groups (identified by those starting with `pg_` when granting permissions)
+
 ## [0.7.0] - 2023-12-24
 
 ### Added